requirements.txt
PySide6<=6.8.1
minecraft-launcher-lib==7.1
psutil==7.0.0
Pillow==11.3.0
